refactor(core): split database modules and tests
